example-react-app/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
docopt==0.6.2
finnhub-python==2.4.2
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pipreqs==0.4.10
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scipy==1.7.0
six==1.16.0
urllib3==1.26.6
yarg==0.1.9
